alluxio map file larger than 2G

+++ core/client/fs/src/main/java/alluxio/client/file/FileInStreamEmpty.java
@@ -78,52 +78,52 @@ public class FileInStreamEmpty extends FileInStream {
 
 
   @Override
-  public int readByte(int pos) throws IOException {
+  public int readByte(long pos) throws IOException {
     throw new IndexOutOfBoundsException();
 
   }
 
   @Override
-  public boolean readBool(int pos) throws IOException {
+  public boolean readBool(long pos) throws IOException {
     throw new IndexOutOfBoundsException();
 
   }
 
   @Override
-  public int readShort(int pos) throws IOException {
+  public int readShort(long pos) throws IOException {
     throw new IndexOutOfBoundsException();
 
   }
 
   @Override
-  public int readInt(int pos) throws IOException {
+  public int readInt(long pos) throws IOException {
     throw new IndexOutOfBoundsException();
 
   }
 
   @Override
-  public float readFloat(int pos) throws IOException {
+  public float readFloat(long pos) throws IOException {
     throw new IndexOutOfBoundsException();
 
   }
 
   @Override
-  public long readLong(int pos) throws IOException {
+  public long readLong(long pos) throws IOException {
     throw new IndexOutOfBoundsException();
 
   }
 
   @Override
-  public double readDouble(int pos) throws IOException {
+  public double readDouble(long pos) throws IOException {
     throw new IndexOutOfBoundsException();
   }
 
   @Override
-  public String readString(int pos) throws IOException {
+  public String readString(long pos) throws IOException {
     throw new IndexOutOfBoundsException();
   }
   @Override
-  public void readBytes(byte[] bytes, int pos) throws IOException {
+  public void readBytes(byte[] bytes, long pos) throws IOException {
     if(bytes.length!=0){
       throw new IndexOutOfBoundsException();
     }
